requirements.txt
fastapi==0.115.0
langchain==0.3.0
langchain_community==0.3.0
langchain_google_genai==2.0.0
protobuf==5.28.2
pydantic==2.9.2
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.35.0

